feat(input): Split the `file` from the `sql`
